pyproject.toml
[tool.poetry]
name = "sake-image-retrieval"
version = "0.1.0"
description = ""
authors = ["gengen"]
readme = "README.md"
packages = [{include = "sake_image_retrieval"}]

[tool.poetry.dependencies]
python = "^3.10"
pandas = "^2.0.3"
matplotlib = "^3.7.2"
seaborn = "^0.12.2"
japanize-matplotlib = "^1.1.3"
torch = "2.0.0"
pillow = "^10.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
